w13/G2/demo/requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.1
sqlparse==0.4.4
